+++ src/main/java/nl/nn/testtool/storage/file/Storage.java
@@ -240,7 +240,7 @@ public class Storage implements nl.nn.testtool.storage.LogStorage {
 //			metadata = storage.getMetadata(numberOfRecords, metadataNames,
 //					searchValues, MetadataExtractor.VALUE_TYPE_GUI);
 //		} catch(StorageException storageException) {
-//			displayError(storageException.getMessage());
+//			displayAndLogError(storageException);
 //		}
 
 	}